Create DocumentView superclass
