+++ lib/nanoc/base/views.rb
@@ -1,4 +1,5 @@
 require_relative 'views/config'
+require_relative 'views/document'
 require_relative 'views/identifiable_collection'
 require_relative 'views/item'
 require_relative 'views/item_collection'
